SDL as a custom audio backend and minor fixes
Problem Statement

Starting from version 5.0, Raylib will support the use of SDL in addition to GLFW. This enables the use of Raylib on some new platforms that were previously unsupported, such as consoles. Porting Raylib to a Nintendo Switch Developer Console was seamlessly possible with the new Platform_Desktop_SDL. However, using raudio in its current form is not possible on these new platforms.

Solution

Miniaudio allows the use of custom backends. To enable audio playback on the new SDL platforms (such as consoles), SDL has been implemented as a custom backend in RAUDIO. By defining #define PLATFORM_SDL in raudio.c, SDL is used as the backend for audio playback. This should enable raudio to provide audio playback on all platforms supported by SDL.

Notes

A new file "backend_sdl.h" has been created in /external. This is largely based on the SDL backend provided by miniaudio. raudio.c has been modified to enable the use of this backend.

This is merely a non-elegant implementation. As it was noted in the wishlist of raylib 5.0, a fundamental revision of raudio.c is being aimed for, and this implementation should be better integrated in that process.

The SDL audio backend has so far only been tested on Windows 11 and a Nintendo Switch Developer Console.

BUGFIXES:

The following small fixes have been made:

1. Implementation of an improved TRACELOG
The output of TRACELOG was error-prone in raudio standalone mode. By integrating <stdarg.h> and implementing a TRACELOG function, correct output in standalone mode has been enabled.

2. SaveFileData and SaveFileText Functions in raudio.c
The functions SaveFileData and SaveFileText in raudio.c were causing compilation warnings due to missing return values. These functions have been updated with return values, resolving the warnings.